Sort stocktake delete into respective folder/files
